scaffolding cleanup
