+++ programs/liquidation/src/state/mod.rs
@@ -1,3 +1,5 @@
 pub mod liquidation;
+pub mod refund_record;
 
 pub use liquidation::*;
+pub use refund_record::*;
